# docs/openapi/orders.yaml
openapi: 3.0.3
info:
  title: orders-service API
  description: |
    Microservicio de gestión de pedidos con trazabilidad completa y métricas de eficiencia.

    **Funcionalidades principales:**
    - Gestión completa del ciclo de vida de pedidos
    - Trazabilidad en tiempo real con MongoDB
    - Métricas de eficiencia de restaurantes y empleados (HU-018)
    - Control de acceso basado en roles (JWT)

    **Arquitectura:**
    - Hexagonal Architecture
    - PostgreSQL para datos transaccionales
    - MongoDB para eventos de trazabilidad
    - Integración con Restaurant Service para validación de autorización
  version: 1.0.0
  contact:
    name: Orders Service Team
    email: orders@pragma.com.co
...
servers:
  - url: http://localhost:8082
    description: orders-service local development
  - url: https://orders-api.pragma.com.co
    description: orders-service production
security:
  - bearerAuth: []
paths:
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Create order (CUSTOMER)
      description: |
        Creates a new order with initial status PENDIENTE. Only customers can create orders.
        Business rules:
        - All dishes must belong to the same restaurant
        - Customer must not have another active order (PENDIENTE, EN_PREPARACION, LISTO)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict (active order exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Orders]
      summary: List orders by status and restaurant (EMPLOYEE/OWNER)
      description: |
        HU-012: Listar Pedidos (EMPLOYEE)
        Lists orders with strict access control:
        - EMPLOYEE: Can only see orders from their assigned restaurant
        - OWNER: Can see orders from their owned restaurants
        - System validates employee-restaurant assignment via Users Service

        Security Implementation:
        - Validates user role and restaurant assignment
        - Returns HTTP 403 if trying to access unauthorized restaurant orders
        - Supports filtering by status and pagination
      parameters:
        - in: query
          name: status
          description: Filter orders by status
          schema: { type: string, enum: [PENDIENTE, EN_PREPARACION, LISTO, ENTREGADO, CANCELADO] }
        - in: query
          name: restaurantId
          description: Restaurant ID to filter orders (validated against user permissions)
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageResponse'
        '400':
          description: Bad request (missing or invalid parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': { description: Unauthorized (missing or invalid token) }
        '403':
          description: Forbidden (user is not employee/owner of the specified restaurant)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/my-orders:
    get:
      tags: [Orders]
      summary: Get my orders (CUSTOMER only)
      description: |
        Retrieves all orders for the authenticated customer with optional status filtering.
        Only customers can access this endpoint and only see their own orders.

        **Access Control:**
        - **CUSTOMER**: Can view only their own orders

        **Features:**
        - Optional status filtering
        - Pagination support
        - Sorted by creation date (newest first)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDIENTE, EN_PREPARACION, LISTO, ENTREGADO, CANCELADO]
          description: Optional filter by order status
          example: "PENDIENTE"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-based)
          example: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of orders per page
          example: 10
      responses:
        '200':
          description: OK - Customer orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageResponse'
              example:
                content:
                  - id: 25
                    customerId: 4
                    restaurantId: 1
                    status: "ENTREGADO"
                    createdAt: "2024-01-15T10:30:00"
                    plates:
                      - id: 1
                        name: "Hamburguesa Clásica"
                        quantity: 2
                totalPages: 1
                totalElements: 1
                size: 10
                number: 0
        '400':
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401': { description: Unauthorized (missing or invalid token) }
        '403': { description: Forbidden (only customers can access this endpoint) }
...
  /api/v1/orders/{id}/cancel:
    put:
      tags: [Orders]
      summary: Cancel order (only PENDIENTE)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400': { description: Bad request }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
        '404': { description: Not found }
  /api/v1/orders/{id}/assign:
    put:
      tags: [Orders]
      summary: Assign yourself to order (EMPLOYEE)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400': { description: Bad request }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
        '404': { description: Not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido del servicio de autenticación.

        **Formato:** `Authorization: Bearer <token>`

        **Roles soportados:**
        - `CUSTOMER`: Clientes del sistema
        - `EMPLOYEE`: Empleados de restaurantes
        - `OWNER`: Propietarios de restaurantes
        - `ADMIN`: Administradores del sistema
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        path: { type: string }
        message: { type: string }
        code: { type: string }
    OrderItemRequest:
      type: object
      required: [dishId, quantity]
      properties:
        dishId: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1 }
    OrderCreateRequest:
      type: object
      required: [customerId, restaurantId, items]
      properties:
        customerId: { type: integer, format: int64 }
        restaurantId: { type: integer, format: int64 }
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemResponse:
      type: object
      properties:
        dishId: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1 }
    OrderResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        customerId: { type: integer, format: int64 }
        restaurantId: { type: integer, format: int64 }
        status: { type: string, enum: [PENDIENTE, EN_PREPARACION, LISTO, ENTREGADO, CANCELADO] }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
    OrderPageResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/OrderResponse' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
...